fixing font, dark/light mode; adding modelview
