+++ NanoChallenge3/View/PortionResultView.swift
@@ -22,25 +22,19 @@ struct PortionResultView: View {
             VStack{
                 VStack(alignment: .leading, spacing: 20) {
                     Text("Recipe portion")
-                    .font(
-                    Font.custom("SF Pro", size: 20)
-                    .weight(.bold)
-                    )
+                        .fontWeight(.bold)
+                        .font(.system(size: 20))
                     .frame(maxWidth: .infinity, alignment: .leading)
                     Text("Enter the portions that the recipe yields")
-                    .font(
-                    Font.custom("SF Pro", size: 16)
-                    .weight(.light)
-                    )
+                        .fontWeight(.light)
+                        .font(.system(size: 16))
                     .frame(maxWidth: .infinity, alignment: .leading)
                     HStack{
                         HStack(alignment: .center, spacing: 8) {
                             // Headline
                             Text("600")
-                            .font(
-                            Font.custom("SF Pro", size: 63)
-                            .weight(.bold)
-                            )
+                                .fontWeight(.bold)
+                                .font(.system(size: 63))
                         }
                         .padding(.horizontal, 8)
                         .padding(.vertical, 12)
@@ -51,11 +45,9 @@ struct PortionResultView: View {
                         HStack(alignment: .center) {
                             Spacer()
                             Text("gr")
-                            .font(
-                            Font.custom("SF Pro", size: 20)
-                            .weight(.semibold)
-                            )
-                            .foregroundColor(Color(red: 0.24, green: 0.24, blue: 0.26))
+                                .fontWeight(.semibold)
+                                .font(.system(size: 20))
+//                            .foregroundColor(Color(red: 0.24, green: 0.24, blue: 0.26))
                             .frame(maxWidth: .infinity, alignment: .topLeading)
                             Spacer()
                         }
@@ -72,16 +64,12 @@ struct PortionResultView: View {
                 
                 VStack(alignment: .leading, spacing: 20) {
                     Text("Target portion")
-                    .font(
-                    Font.custom("SF Pro", size: 20)
-                    .weight(.bold)
-                    )
+                        .fontWeight(.bold)
+                        .font(.system(size: 20))
                     .frame(maxWidth: .infinity, alignment: .leading)
                     Text("Your target portion want to make")
-                    .font(
-                    Font.custom("SF Pro", size: 16)
-                    .weight(.light)
-                    )
+                        .fontWeight(.light)
+                        .font(.system(size: 16))
                     .frame(maxWidth: .infinity, alignment: .leading)
                     
                     // buat input
@@ -92,10 +80,8 @@ struct PortionResultView: View {
                             TextField("", text: $recipePortionInput)
                                 .focused($focusedField, equals: .dec)
                                 .numbersOnly($recipePortionInput)
-                            .font(
-                            Font.custom("SF Pro", size: 63)
-                            .weight(.bold)
-                            )
+                                .fontWeight(.bold)
+                                .font(.system(size: 63))
                         }
                         .padding(.horizontal, 8)
                         .padding(.vertical, 12)
@@ -110,11 +96,9 @@ struct PortionResultView: View {
                         HStack(alignment: .center) {
                             Spacer()
                             Text("gr")
-                            .font(
-                            Font.custom("SF Pro", size: 20)
-                            .weight(.semibold)
-                            )
-                            .foregroundColor(Color(red: 0.24, green: 0.24, blue: 0.26))
+                                .fontWeight(.semibold)
+                                .font(.system(size: 20))
+//                            .foregroundColor(Color(red: 0.24, green: 0.24, blue: 0.26))
                             .frame(maxWidth: .infinity, alignment: .topLeading)
                             Spacer()
                         }
@@ -133,19 +117,15 @@ struct PortionResultView: View {
                     VStack {
                         // Body/Body Semibold 2
                         Text("Ingredients information")
-                          .font(
-                            Font.custom("SF Pro", size: 20)
-                              .weight(.semibold)
-                          )
+                            .fontWeight(.semibold)
+                            .font(.system(size: 20))
                           .foregroundColor(Color(red: 0, green: 0.54, blue: 0.82))
                           .frame(maxWidth: .infinity, alignment: .leading)
                         
                         // Body/Body 3
                         Text("Composition needed to fulfill your target")
-                          .font(
-                            Font.custom("SF Pro", size: 16)
-                              .weight(.light)
-                          )
+                            .fontWeight(.light)
+                            .font(.system(size: 16))
                           .foregroundColor(Color(red: 0.2, green: 0.72, blue: 1))
                           .frame(maxWidth: .infinity, alignment: .leading)
                         
@@ -153,10 +133,8 @@ struct PortionResultView: View {
                             HStack(alignment: .center, spacing: 12) {
                                 HStack(alignment: .center, spacing: 20) {
                                     Text("1000")
-                                    .font(
-                                    Font.custom("SF Pro", size: 16)
-                                    .weight(.semibold)
-                                    )
+                                        .fontWeight(.semibold)
+                                        .font(.system(size: 20))
                                     .foregroundColor(Color(red: 0, green: 0.54, blue: 0.82))
 
                                     .frame(maxWidth: .infinity, alignment: .topLeading)
@@ -174,10 +152,8 @@ struct PortionResultView: View {
                                 HStack(alignment: .center, spacing: 20) {
                                     // Body/Body Semibold 3
                                     Text("gr")
-                                      .font(
-                                        Font.custom("SF Pro", size: 16)
-                                          .weight(.semibold)
-                                      )
+                                        .fontWeight(.semibold)
+                                        .font(.system(size: 16))
                                       .foregroundColor(Color(red: 0, green: 0.54, blue: 0.82))
                                       .frame(maxWidth: .infinity, alignment: .topLeading)
                                 }
@@ -194,10 +170,8 @@ struct PortionResultView: View {
                                 HStack(alignment: .center, spacing: 20) {
                                     // Body/Body Semibold 3
                                     Text("salted butter")
-                                      .font(
-                                        Font.custom("SF Pro", size: 16)
-                                          .weight(.semibold)
-                                      )
+                                        .fontWeight(.semibold)
+                                        .font(.system(size: 16))
                                       .foregroundColor(Color(red: 0, green: 0.54, blue: 0.82))
                                       .frame(maxWidth: .infinity, alignment: .topLeading)
                                 }
@@ -222,10 +196,8 @@ struct PortionResultView: View {
                     } label: {
                         HStack(alignment: .center, spacing: 4) {
                             Text("Calculate")
-                            .font(
-                            Font.custom("SF Pro", size: 20)
-                            .weight(.semibold)
-                            )
+                                .fontWeight(.semibold)
+                                .font(.system(size: 20))
                             .foregroundColor(.white)
                         }
                         .padding(.horizontal, 20)
